Avoid allocation of SearchResult instance upon looking up an exact value where from a ConcurrentRadixTree.
By avoiding this allocation, looking up a large chunk of values can be done with less GC overhead. Escape analysis on the JVM is unfortunately rather limited such that this allocation is not properly avoided by the JIT in many cases. If this allocation ends up on the hot path, its collection can have a visible impact on throughput.