Introduce NodeList abstraction to avoid list allocation when reading decendent nodes.
Unfortunately, the AbstractReferenceArray API and List API are incompatible due to the conflicting signatures of set(int,Object) where one returns a boolean and another declares void. To avoid this conflict, an API is introduced that implements the minimal required functionality but allows a single object to represent any interaction with child nodes. This way, non-leaf nodes do no longer require an additional field to carry the list wrapper, reducing the memory consumption of a tree with many non-leaf nodes.